+++ src/sql/cmd/profile.py
@@ -1,20 +1,25 @@
 from sql import inspect
 from sql.cmd.cmd_utils import CmdParser
+from sql.util import expand_args, is_rendering_required
 
 
-def profile(others):
+def profile(others, user_ns):
     """
     Implementation of `%sqlcmd profile`
     This function takes in a string containing command line arguments,
     parses them to extract the name of the table, the schema, and the output location.
     It then retrieves statistical information about the specified table and either
     returns the report or writes it to the specified location.
+    It also uses the kernel namespace for expanding arguments declared as variables.
 
 
     Parameters
     ----------
     others : str,
         A string containing the command line arguments.
+
+    user_ns : dict,
+        User namespace of IPython kernel
 
     Returns
     -------
@@ -31,6 +36,8 @@ def profile(others):
     )
 
     args = parser.parse_args(others)
+    if is_rendering_required(" ".join(others)):
+        expand_args(args, user_ns)
 
     report = inspect.get_table_statistics(schema=args.schema, name=args.table)