+++ src/sql/command.py
@@ -90,6 +90,9 @@ class SQLCommand:
             self.parsed["connection"] = self.args.line[0]
 
         if self.args.with_:
+            self.args.with_ = [
+                Template(item).render(user_ns) for item in self.args.with_
+            ]
             final = store.render(self.parsed["sql"], with_=self.args.with_)
             self.parsed["sql"] = str(final)