+++ src/tests/test_magic_cmd.py
@@ -96,6 +96,24 @@ INSERT INTO schema1.table1 VALUES (2, 'two');
 
 
 @pytest.mark.parametrize(
+    "cmd, cols, table_name",
+    [
+        [
+            "%sqlcmd columns -t {{table}}",
+            ["first_name", "last_name", "year_of_death"],
+            "author",
+        ],
+        ["%sqlcmd columns -t {{table}}", ["first", "second"], "table with spaces"],
+        ["%sqlcmd columns -t {{table}}", ["first", "second"], "table with spaces"],
+    ],
+)
+def test_columns_with_variable_substitution(ip, cmd, cols, table_name):
+    ip.user_global_ns["table"] = table_name
+    out = ip.run_cell(cmd).result._repr_html_()
+    assert all(col in out for col in cols)
+
+
+@pytest.mark.parametrize(
     "cell, error_message",
     [
         [
@@ -205,6 +223,23 @@ ATTACH DATABASE 'my.db' AS some_schema
     assert "numbers" in out
 
 
+def test_tables_with_schema_variable_substitution(ip, tmp_empty):
+    conn = SQLAlchemyConnection(engine=create_engine("sqlite:///my.db"))
+    conn.execute("CREATE TABLE numbers (some_number FLOAT)")
+
+    ip.user_global_ns["schema"] = "some_schema"
+
+    ip.run_cell(
+        """%%sql
+ATTACH DATABASE 'my.db' AS {{schema}}
+"""
+    )
+
+    out = ip.run_cell("%sqlcmd tables --schema {{schema}}").result._repr_html_()
+
+    assert "numbers" in out
+
+
 @pytest.mark.parametrize(
     "cmd, cols",
     [
@@ -234,6 +269,28 @@ def test_columns_with_schema(ip, tmp_empty, arguments):
     ip.run_cell(
         """%%sql
 ATTACH DATABASE 'my.db' AS some_schema
+"""
+    )
+
+    out = ip.run_cell(f"%sqlcmd columns {arguments}").result._repr_html_()
+
+    assert "some_number" in out
+
+
+@pytest.mark.parametrize(
+    "arguments",
+    ["--table {{table}} --schema {{schema}}", "--table {{schema}}.{{table}}"],
+)
+def test_columns_with_schema_variable_substitution(ip, tmp_empty, arguments):
+    conn = SQLAlchemyConnection(engine=create_engine("sqlite:///my.db"))
+    conn.execute("CREATE TABLE numbers (some_number FLOAT)")
+
+    ip.user_global_ns["table"] = "numbers"
+    ip.user_global_ns["schema"] = "some_schema"
+
+    ip.run_cell(
+        """%%sql
+ATTACH DATABASE 'my.db' AS {{schema}}
 """
     )
 
@@ -277,6 +334,65 @@ def test_table_profile(ip_with_connections, tmp_empty, conn):
     }
 
     out = ip_with_connections.run_cell("%sqlcmd profile -t numbers").result
+
+    stats_table = out._table
+
+    assert len(stats_table.rows) == len(expected)
+
+    for row in stats_table:
+        profile_metric = _get_row_string(row, " ")
+        rating = _get_row_string(row, "rating")
+        price = _get_row_string(row, "price")
+        number = _get_row_string(row, "number")
+        word = _get_row_string(row, "word")
+
+        assert profile_metric in expected
+        assert rating == str(expected[profile_metric][0])
+        assert price == str(expected[profile_metric][1])
+        assert number == str(expected[profile_metric][2])
+        assert word == str(expected[profile_metric][3])
+
+    # Test sticky column style was injected
+    assert "position: sticky;" in out._table_html
+
+
+@pytest.mark.parametrize(
+    "conn",
+    [
+        ("sqlite_sqlalchemy"),
+        ("sqlite_dbapi"),
+    ],
+)
+def test_table_profile_with_substitution(ip_with_connections, tmp_empty, conn):
+    ip_with_connections.run_cell(f"%sql {conn}")
+    ip_with_connections.run_cell(
+        """
+    %%sql
+    CREATE TABLE numbers (rating float, price float, number int, word varchar(50));
+    INSERT INTO numbers VALUES (14.44, 2.48, 82, 'a');
+    INSERT INTO numbers VALUES (13.13, 1.50, 93, 'b');
+    INSERT INTO numbers VALUES (12.59, 0.20, 98, 'a');
+    INSERT INTO numbers VALUES (11.54, 0.41, 89, 'a');
+    INSERT INTO numbers VALUES (10.532, 0.1, 88, 'c');
+    INSERT INTO numbers VALUES (11.5, 0.2, 84, '   ');
+    INSERT INTO numbers VALUES (11.1, 0.3, 90, 'a');
+    INSERT INTO numbers VALUES (12.9, 0.31, 86, '');
+    """
+    )
+
+    expected = {
+        "count": [8, 8, 8, 8],
+        "mean": ["12.2165", "0.6875", "88.7500", math.nan],
+        "min": [10.532, 0.1, 82, math.nan],
+        "max": [14.44, 2.48, 98, math.nan],
+        "unique": [8, 7, 8, 5],
+        "freq": [math.nan, math.nan, math.nan, 4],
+        "top": [math.nan, math.nan, math.nan, "a"],
+    }
+
+    ip_with_connections.user_global_ns["table"] = "numbers"
+
+    out = ip_with_connections.run_cell("%sqlcmd profile -t {{table}}").result
 
     stats_table = out._table
 
@@ -397,6 +513,59 @@ def test_table_schema_profile(ip, tmp_empty, arguments):
         "top": [math.nan],
     }
 
+    out = ip.run_cell(f"%sqlcmd profile {arguments}").result
+
+    stats_table = out._table
+
+    for row in stats_table:
+        profile_metric = _get_row_string(row, " ")
+
+        cell = row.get_string(fields=["n"], border=False, header=False).strip()
+
+        if profile_metric in expected:
+            assert cell == str(expected[profile_metric][0])
+
+
+@pytest.mark.parametrize(
+    "arguments",
+    ["--table {{table}} --schema {{schema}}", "--table {{schema}}.{{table}}"],
+)
+def test_table_schema_profile_with_substitution(ip, tmp_empty, arguments):
+    ip.run_cell("%sql sqlite:///a.db")
+    ip.run_cell("%sql CREATE TABLE t (n FLOAT)")
+    ip.run_cell("%sql INSERT INTO t VALUES (1)")
+    ip.run_cell("%sql INSERT INTO t VALUES (2)")
+    ip.run_cell("%sql INSERT INTO t VALUES (3)")
+    ip.run_cell("%sql --close sqlite:///a.db")
+
+    ip.run_cell("%sql sqlite:///b.db")
+    ip.run_cell("%sql CREATE TABLE t (n FLOAT)")
+    ip.run_cell("%sql INSERT INTO t VALUES (11)")
+    ip.run_cell("%sql INSERT INTO t VALUES (22)")
+    ip.run_cell("%sql INSERT INTO t VALUES (33)")
+    ip.run_cell("%sql --close sqlite:///b.db")
+
+    ip.run_cell(
+        """
+    %%sql sqlite://
+    ATTACH DATABASE 'a.db' AS a_schema;
+    ATTACH DATABASE 'b.db' AS b_schema;
+    """
+    )
+
+    expected = {
+        "count": ["3"],
+        "mean": ["22.0000"],
+        "min": ["11.0"],
+        "max": ["33.0"],
+        "std": ["11.0000"],
+        "unique": ["3"],
+        "freq": [math.nan],
+        "top": [math.nan],
+    }
+
+    ip.user_global_ns["table"] = "t"
+    ip.user_global_ns["schema"] = "b_schema"
     out = ip.run_cell(f"%sqlcmd profile {arguments}").result
 
     stats_table = out._table
@@ -549,6 +718,35 @@ def test_table_profile_store(ip_with_connections, tmp_empty, conn, report_fname)
 
 
 @pytest.mark.parametrize(
+    "conn, report_fname",
+    [
+        ("sqlite_sqlalchemy", "test_report.html"),
+        ("sqlite_dbapi", "test_report_dbapi.html"),
+    ],
+)
+def test_table_profile_store_with_substitution(
+    ip_with_connections, tmp_empty, conn, report_fname
+):
+    ip_with_connections.run_cell(
+        f"""
+    %%sql {conn}
+    CREATE TABLE test_store (rating, price, number, symbol);
+    INSERT INTO test_store VALUES (14.44, 2.48, 82, 'a');
+    INSERT INTO test_store VALUES (13.13, 1.50, 93, 'b');
+    INSERT INTO test_store VALUES (12.59, 0.20, 98, 'a');
+    INSERT INTO test_store VALUES (11.54, 0.41, 89, 'a');
+    """
+    )
+    ip_with_connections.user_global_ns["table"] = "test_store"
+    ip_with_connections.user_global_ns["output"] = report_fname
+
+    ip_with_connections.run_cell("%sqlcmd profile -t {{table}} --output {{output}}")
+
+    report = Path(report_fname)
+    assert report.is_file()
+
+
+@pytest.mark.parametrize(
     "cell, error_message",
     [
         [
@@ -606,6 +804,27 @@ def test_failing_test_with_schema(ip_empty, sample_schema_with_table, arguments)
 )
 def test_passing_test_with_schema(ip_empty, sample_schema_with_table, arguments):
     out = ip_empty.run_cell(f"%sqlcmd test {arguments} --column x --less-than 3").result
+    assert out is True
+
+
+@pytest.mark.parametrize(
+    "arguments",
+    ["--table {{schema}}.{{table}}", "--table {{table}} --schema {{schema}}"],
+)
+def test_test_with_schema_variable_substitution(
+    ip_empty, sample_schema_with_table, arguments
+):
+    ip_empty.user_global_ns["table"] = "table1"
+    ip_empty.user_global_ns["schema"] = "schema1"
+    out = ip_empty.run_cell(f"%sqlcmd test {arguments} --column x --less-than 3").result
+    assert out is True
+
+
+def test_test_column_variable_substitution(ip_empty, sample_schema_with_table):
+    ip_empty.user_global_ns["column"] = "x"
+    out = ip_empty.run_cell(
+        "%sqlcmd test --table schema1.table1 --column {{column}} --less-than 3"
+    ).result
     assert out is True
 
 
@@ -712,6 +931,15 @@ def test_delete_saved_key(ip_snippets, arg):
     assert "high_price_a" not in stored_snippets
 
 
+def test_delete_saved_key_with_variable_substitution(ip_snippets):
+    ip_snippets.user_global_ns["snippet_name"] = "high_price_a"
+    out = ip_snippets.run_cell("%sqlcmd snippets --delete {{snippet_name}}").result
+    assert "high_price_a has been deleted.\n" in out
+    stored_snippets = out[out.find("Stored snippets") + len("Stored snippets: ") :]
+    assert "high_price, high_price_b" in stored_snippets
+    assert "high_price_a" not in stored_snippets
+
+
 @pytest.mark.parametrize("arg", ["--delete-force", "-D"])
 def test_force_delete(ip_snippets, arg):
     out = ip_snippets.run_cell(f"%sqlcmd snippets {arg} high_price").result
@@ -724,9 +952,32 @@ def test_force_delete(ip_snippets, arg):
     assert "high_price," not in stored_snippets
 
 
+def test_force_delete_with_variable_substitution(ip_snippets):
+    ip_snippets.user_global_ns["snippet_name"] = "high_price"
+    out = ip_snippets.run_cell(
+        "%sqlcmd snippets --delete-force {{snippet_name}}"
+    ).result
+    assert (
+        "high_price has been deleted.\nhigh_price_a, "
+        "high_price_b depend on high_price\n" in out
+    )
+    stored_snippets = out[out.find("Stored snippets") + len("Stored snippets: ") :]
+    assert "high_price_a, high_price_b" in stored_snippets
+    assert "high_price," not in stored_snippets
+
+
 @pytest.mark.parametrize("arg", ["--delete-force-all", "-A"])
 def test_force_delete_all(ip_snippets, arg):
     out = ip_snippets.run_cell(f"%sqlcmd snippets {arg} high_price").result
+    assert "high_price_a, high_price_b, high_price has been deleted" in out
+    assert "There are no stored snippets" in out
+
+
+def test_force_delete_all_with_variable_substitution(ip_snippets):
+    ip_snippets.user_global_ns["snippet_name"] = "high_price"
+    out = ip_snippets.run_cell(
+        "%sqlcmd snippets --delete-force-all {{snippet_name}}"
+    ).result
     assert "high_price_a, high_price_b, high_price has been deleted" in out
     assert "There are no stored snippets" in out
 
@@ -796,12 +1047,39 @@ def test_delete_snippet_when_dependency_force_deleted(ip_snippets, arg):
     assert "high_price_a has been deleted.\nStored snippets: high_price_b" in out
 
 
+def test_view_snippet_variable_substitution(ip_snippets):
+    ip_snippets.user_global_ns["snippet_name"] = "test_snippet"
+    ip_snippets.run_cell(
+        """%%sql --save {{snippet_name}} --no-execute
+SELECT * FROM "test_store" WHERE price >= 1.50
+"""
+    )
+
+    out = ip_snippets.run_cell("%sqlcmd snippets {{snippet_name}}").result
+    assert 'SELECT * FROM "test_store" WHERE price >= 1.50' in out
+
+
 @pytest.mark.parametrize(
     "arguments", ["--table schema1.table1", "--table table1 --schema schema1"]
 )
 def test_explore_with_schema(ip_empty, sample_schema_with_table, arguments):
     expected_rows = ['"x": 1', '"y": "one"', '"x": 2', '"y": "two"']
 
+    out = ip_empty.run_cell(f"%sqlcmd explore {arguments}").result
+    assert isinstance(out, TableWidget)
+    assert [row in out._repr_html_() for row in expected_rows]
+
+
+@pytest.mark.parametrize(
+    "arguments",
+    ["--table {{schema}}.{{table}}", "--table {{table}} --schema {{schema}}"],
+)
+def test_explore_with_schema_variable_substitution(
+    ip_empty, sample_schema_with_table, arguments
+):
+    expected_rows = ['"x": 1', '"y": "one"', '"x": 2', '"y": "two"']
+    ip_empty.user_global_ns["table"] = "table1"
+    ip_empty.user_global_ns["schema"] = "schema1"
     out = ip_empty.run_cell(f"%sqlcmd explore {arguments}").result
     assert isinstance(out, TableWidget)
     assert [row in out._repr_html_() for row in expected_rows]